feature: Break down Simulation() into Simulation() and StateVectorSimulation()
Simulation() contains basic properties and methods that share among different simulation methods.

StateVectorSimulation() contains those which are specific to statevector.
